🐛 Fix: useWindowWidth 훅 반환값 변경 및 그에 따른 이름 변경, Paginator visblePages 조건 변경
