--- src/shared/hooks/useWindowWidth.tsx
@@ -1,21 +1,0 @@
-import { useEffect, useState } from "react";
-
-// 현재 창의 너비를 반환하는 커스텀 훅
-const useWindowWidth = () => {
-  const [width, setWidth] = useState<number | null>(null);
-
-  useEffect(() => {
-    const handleResize = () => {
-      setWidth(window.innerWidth);
-    };
-    handleResize();
-    window.addEventListener("resize", handleResize);
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
-
-  return width;
-};
-
-export default useWindowWidth;
